pyproject.toml
[tool.poetry]
name = "icortex"
version = "0.1.3"
description = "Jupyter kernel that can generate Python code from natural language prompts"
authors = ["TextCortex Team <onur@textcortex.com>"]
license = "Apache"
readme = "README.md"
repository = "https://github.com/textcortex/icortex"
homepage = "https://icortex.ai/"
keywords = [
    "copilot",
    "tabnine",
    "codex",
    "openai",
    "code generation",
    "code completion",
    "generate code",
    "natural language to python",
    "icortex",
    "textcortex",
    "gpt-3",
    "codegen",
    "polycoder",
    "jupyter notebook",
    "smart jupyter notebook",
    "jupyter notebook kernel",
    "icortex kernel",
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry.scripts]
icortex = "icortex.cli:main"

[tool.poetry.dependencies]
python = ">=3.7.1,<4"
toml = "^0.10.2"
importlib-metadata = ">=4.0.0"
# ipykernel = "^5.5.5"
requests = "^2.28.1"
urllib3 = "^1.26.12"
ipykernel = "^6.16.0"
ipython = ">=7.0.0"
entrypoints = "^0.4"
ipywidgets = "^8.0.2"
jupyter-client = "^7.4.2"
jupyter-console = "^6.4.4"
jupyter-core = "^4.11.1"
jupyterlab-widgets = "^3.0.3"
Pygments = "^2.13.0"
openai = { version = "^0.23.1", optional = true }
onnxruntime = { version = "^1.12.1", optional = true }
onnx = { version = "^1.12.0", optional = true }
optimum = { version = "^1.4.0", optional = true }
transformers = { version = "^4.23.1", optional = true }
torch = { version = "^1.12.1", optional = true }
black = "^22.10.0"

[tool.poetry.extras]
openai = ["openai"]
huggingface = ["onnxruntime", "onnx", "optimum", "transformers", "torch"]

[tool.poetry.group.dev.dependencies]
Sphinx = "^5.3.0"
sphinxcontrib-video = "^0.0.1.dev3"
sphinx-copybutton = "^0.5.0"
furo = "^2022.9.29"
pytest = "^7.1.3"
python-dotenv = "^0.21.0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
